Protecting-VetControllers

+++ Veterinarian.Api/Controllers/VetsController.cs
@@ -1,55 +1,109 @@
 ﻿using FluentValidation;
 using FluentValidation.Results;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using System.Threading;
+using Veterinaria.Domain.Entities.Abstracts;
 using Veterinaria.Domain.Entities.Users;
 using Veterinarian.Application.AuthServices;
 using Veterinarian.Application.Users;
+using Veterinarian.Application.UserServices;
 using Veterinarian.Application.Vets;
 
 namespace Veterinarian.Api.Controllers
 {
+    [Authorize]
     [ApiController]
-    [Route("Api/Vets")]
+    [Route("api/vets")]
     public class VetsController : ControllerBase
     {
         private readonly IVetServices _vetServices;
         private readonly IUserManagerServices _userManagerServices;
         private readonly RoleManager<IdentityRole> _roleManager;
+        private readonly IUserContext _userContext;
 
         public VetsController(IVetServices vetServices,
             IUserManagerServices userManagerServices,
-            RoleManager<IdentityRole> roleManager)
+            RoleManager<IdentityRole> roleManager,
+            IUserContext userContext)
         {
             _vetServices = vetServices;
             _userManagerServices = userManagerServices;
             _roleManager = roleManager;
+            _userContext = userContext;
         }
 
-        [HttpGet("GetVets")]
-        public async Task<IActionResult> GetVets()
+        [Authorize(Roles = $"{Role.AuxiliaryMember},{Role.Admin}")]
+        [HttpGet("getVets")]
+        public async Task<IActionResult> GetVets(CancellationToken cancellationToken)
         {
+
+            string? CurrentUserId = await _userContext.GetUserIdAsync(cancellationToken);
+
+            if (string.IsNullOrWhiteSpace(CurrentUserId))
+            {
+                return Problem(
+                    detail: "Unauthorized",
+                    statusCode:StatusCodes.Status401Unauthorized);
+            }
+
+
             var vets = await _vetServices.GetAllAsync();
             return Ok(vets.Value);
         }
 
-        [HttpGet("GetVet/{id}")]
-        public async Task<IActionResult> GetVetById(Guid id)
+        [Authorize(Roles =$"{Role.Admin},{Role.AuxiliaryMember},{Role.VetMember}")]
+        [HttpGet("getVet/{id}")]
+        public async Task<IActionResult> GetVetById(Guid id, CancellationToken cancellationToken)
         {
+
+            string? CurrentUserId = await _userContext.GetUserIdAsync(cancellationToken);
+
+            if (string.IsNullOrWhiteSpace(CurrentUserId))
+            {
+                return Problem(
+                    detail: "Unauthorized",
+                    statusCode: StatusCodes.Status401Unauthorized);
+            }
+
+
             var vet = await _vetServices.GetByIdAsync(id);
             return vet.IsSuccess ? Ok(vet.Value) : NotFound(vet.Error);
         }
-        [HttpGet("SearchVet")]
-        public async Task<IActionResult> SerchVets([FromQuery] string? search, int page = 1, int pageSize = 5)
+
+        [Authorize(Roles = $"{Role.Admin},{Role.AuxiliaryMember}")]
+        [HttpGet("searchVet")]
+        public async Task<IActionResult> SerchVets([FromQuery] string? search,CancellationToken cancellationToken, int page = 1, int pageSize = 5)
         {
+
+            string? CurrentUserId = await _userContext.GetUserIdAsync(cancellationToken);
+
+            if (string.IsNullOrWhiteSpace(CurrentUserId))
+            {
+                return Problem(
+                    detail: "Unauthorized",
+                    statusCode: StatusCodes.Status401Unauthorized);
+            }
+
+
             var vets = await _vetServices.SearchVetAsync(search,page,pageSize);
             return Ok(vets.Value);
         }
 
+        [Authorize(Roles = $"{Role.Admin},{Role.AuxiliaryMember}")]
         [HttpPost]
         public async Task<IActionResult> CreateVet([FromBody] VetRequest request,
-            IValidator<VetRequest> validator)
+            IValidator<VetRequest> validator, CancellationToken cancellationToken)
         {
+            string? CurrentUserId = await _userContext.GetUserIdAsync(cancellationToken);
+
+            if (string.IsNullOrWhiteSpace(CurrentUserId))
+            {
+                return Problem(
+                    detail: "Unauthorized",
+                    statusCode: StatusCodes.Status401Unauthorized);
+            }
             ValidationResult validationResult = await validator.ValidateAsync(request);
 
             if(!validationResult.IsValid)
@@ -113,17 +167,40 @@ namespace Veterinarian.Api.Controllers
             var response = await _vetServices.CreateAndRegisterAsync(request,identityUser);
             return response.IsSuccess ? Ok(response) : BadRequest(response.Error);
         }
+
+        [Authorize(Roles = $"{Role.Admin},{Role.AuxiliaryMember}")]
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteVet(Guid id)
+        public async Task<IActionResult> DeleteVet(Guid id, CancellationToken cancellationToken)
         {
+            string? CurrentUserId = await _userContext.GetUserIdAsync(cancellationToken);
+
+            if (string.IsNullOrWhiteSpace(CurrentUserId))
+            {
+                return Problem(
+                    detail: "Unauthorized",
+                    statusCode: StatusCodes.Status401Unauthorized);
+            }
+
+
             var response = await _vetServices.DeleteAsync(id);
             return response.IsSuccess ? NoContent() : NotFound(response.Error);
         }
 
+        [Authorize(Roles = $"{Role.VetMember}")]
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateVet(Guid id, [FromBody] VetRequest request
-            ,IValidator<VetRequest> validator)
+            ,IValidator<VetRequest> validator,CancellationToken cancellationToken)
         {
+
+            string? CurrentUserId = await _userContext.GetUserIdAsync(cancellationToken);
+
+            if (string.IsNullOrWhiteSpace(CurrentUserId))
+            {
+                return Problem(
+                    detail: "Unauthorized",
+                    statusCode: StatusCodes.Status401Unauthorized);
+            }
+
             ValidationResult validationResult = await validator.ValidateAsync(request);
 
             if (!validationResult.IsValid)
@@ -131,7 +208,7 @@ namespace Veterinarian.Api.Controllers
                 return BadRequest(validationResult.ToDictionary());
             }
 
-            var response = await _vetServices.UpdateAsync(id, request);
+            var response = await _vetServices.UpdateAsync(id, request,CurrentUserId);
             return response.IsSuccess ? NoContent() : NotFound(response.Error);
         }
